refactor: move suppressLogs into standardLog.ts
