+++ packages/log/ts/index.ts
@@ -1,9 +1,8 @@
 export * from './console/index.js'
-export * from './createStandardLog.js'
+export * from './standardLog.js'
 export * from './errors.js'
 export * from './formatter.js'
 export * from './logLevels.js'
 export * from './memory.js'
 export * from './shouldLog.js'
-export * from './suppressLogs.js'
 export * from './types.js'
